fixing stuff

+++ berita/views.py
@@ -16,34 +16,51 @@ from berita.forms import BeritaForm
 
 # Create your views here.
 
-@login_required(login_url="homepage:login")
 def show(request):
-    try:
-        count = request.COOKIES['news_count']
-    except:
-        response = HttpResponseRedirect(reverse("berita:show"))
-        response.set_cookie('news_count', 0)
-        return response
-    form = BeritaForm(request.user)
-    try :
-        context = {
+    if request.user.is_superuser:
+        try:
+            count = request.COOKIES['news_count']
+        except:
+            response = HttpResponseRedirect(reverse("berita:show"))
+            response.set_cookie('news_count', 0)
+            return response
+        form = BeritaForm(request.user)
+        try :
+            context = {
+                'last_news' : request.session['last_news'],
+                'count' : count,
+                'form' : form,
+            }
+        except :
+            context = {
+                'last_news' : 'No Last Seen News Yet',
+                'count' : count,
+                'form' : form,
+            }
+        if request.method == 'POST':
+            form = BeritaForm(request.user, request.POST)
+            if form.is_valid():
+                form.save()
+                return JsonResponse({'message': 'success'})
+        return render(request, "admin.html", context)
+    else:
+        try:
+            count = request.COOKIES['news_count']
+        except:
+            response = HttpResponseRedirect(reverse("berita:show"))
+            response.set_cookie('news_count', 0)
+            return response
+        try :
+            context = {
             'last_news' : request.session['last_news'],
             'count' : count,
-            'form' : form,
-        }
-    except :
-        context = {
+            }
+        except :
+            context = {
             'last_news' : 'No Last Seen News Yet',
             'count' : count,
-            'form' : form,
-        }
-    if request.method == 'POST':
-        form = BeritaForm(request.user, request.POST)
-        if form.is_valid():
-            form.save()
-            return JsonResponse({'message': 'success'})
-    return render(request, "admin.html", context)
-
+            }
+        return render(request, "user.html", context)
 
 def json_all(request):
     data = Berita.objects.all()
@@ -51,15 +68,21 @@ def json_all(request):
 
 @login_required(login_url="homepage:login")
 def delete(request, pk):
-    Berita.objects.filter(pk=pk).delete()
-    return JsonResponse({'message': 'success'}) 
+    if request.user.is_superuser:
+        Berita.objects.filter(pk=pk).delete()
+        return JsonResponse({'message': 'success'})
+    else:
+        return render(request, "forbidden.html")   
 
 @login_required(login_url="homepage:login")
 def add(request):
-    if request.method == 'POST':
-        Berita(title=request.POST.get('title'), content=request.POST.get('content'), category=request.POST.get('category'), writer=request.POST.get('writer'), source=request.POST.get('source')).save()
-        return JsonResponse({'message': 'success'})
-    return render(request, "create.html")
+    if request.user.is_superuser:
+        if request.method == 'POST':
+            Berita(title=request.POST.get('title'), content=request.POST.get('content'), category=request.POST.get('category'), writer=request.POST.get('writer'), source=request.POST.get('source')).save()
+            return JsonResponse({'message': 'success'})
+        return render(request, "create.html")
+    else:
+        return render(request, "forbidden.html")
 
 
 def berita(request, pk):
@@ -75,8 +98,11 @@ def berita(request, pk):
 
 @login_required(login_url="homepage:login")
 def delete_filter(request, pk):
-    Berita.objects.filter(pk=pk).delete()
-    return redirect('berita:show')
+    if request.user.is_superuser:
+        Berita.objects.filter(pk=pk).delete()
+        return redirect('berita:show')
+    else:
+        return render(request, "forbidden.html")
